Structure modified. Preparing for player refactor
