+++ app/src/main/java/com/practicum/playlistmaker/ui/player/MediaPlayerController.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker
+package com.practicum.playlistmaker.ui.player
 
 import android.media.MediaPlayer
 import android.os.Handler
