+++ app/src/main/java/com/practicum/playlistmaker/ui/search_history/HistoryRVAdapter.kt
@@ -1,9 +1,11 @@
-package com.practicum.playlistmaker
+package com.practicum.playlistmaker.ui.search_history
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import com.practicum.playlistmaker.R
+import com.practicum.playlistmaker.ui.search.TrackViewHolder
 import com.practicum.playlistmaker.domain.model.Track
 
 class HistoryRVAdapter(
