+++ app/src/main/res/layout/activity_medialibrary.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MediaLibraryActivity">
+    tools:context=".ui.media_library.MediaLibraryActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
